WIP: The blobbiest blob of all blobs that blobbed. (#34)
* Changes:

* Huge change: almost all internal string operations have been moved to a simpler se_string.
* Move more source files to the common se_core library.
* use eastl::vector_set to implement VSet
* use as<type> on some Variant uses.
* Enable OS to store various 'features' dynamically, so that it's not
dependent on TOOLS build flags.
* map/set `keys` function now can return a vector with specific allocator.
* reduce the number of List/PODList usages in the core.
* fix Variant OP_LESS which was returning a string instead of a bool
type.
* some clang-tidying
* Update to upstream c3ea4ea9b76d1b
* Fixups and some Span/vector conversions
* By default ptrcall is disabled since we don't use gdnative.
* Add some still unused files, change underlying types of a few class_db and visual server members
* Fix a few memory leaks.

+++ scene/resources/capsule_shape_2d.cpp
@@ -50,7 +50,7 @@ Vector<Vector2> CapsuleShape2D::_get_points() const {
     return points;
 }
 
-bool CapsuleShape2D::_edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const {
+bool CapsuleShape2D::_edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const {
 
     return Geometry::is_point_in_polygon(p_point, _get_points());
 }
