+++ scene/resources/capsule_shape_2d.h
@@ -46,7 +46,7 @@ protected:
 	static void _bind_methods();
 
 public:
-	bool _edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const override;
+    bool _edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const override;
 
 	void set_height(real_t p_height);
 	real_t get_height() const;
