+++ scene/resources/convex_polygon_shape_2d.cpp
@@ -37,7 +37,7 @@
 
 IMPL_GDCLASS(ConvexPolygonShape2D)
 
-bool ConvexPolygonShape2D::_edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const {
+bool ConvexPolygonShape2D::_edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const {
 
     return Geometry::is_point_in_polygon(p_point, points);
 }
