+++ scene/resources/line_shape_2d.cpp
@@ -36,7 +36,7 @@
 
 IMPL_GDCLASS(LineShape2D)
 
-bool LineShape2D::_edit_is_selected_on_click(const Point2 &p_point, double p_tolerance) const {
+bool LineShape2D::_edit_is_selected_on_click(const Point2 &p_point, float p_tolerance) const {
 
     Vector2 point = get_d() * get_normal();
     Vector2 l[2][2] = { { point - get_normal().tangent() * 100, point + get_normal().tangent() * 100 }, { point, point + get_normal() * 30 } };
